requirements.txt
aiogram
plotly
python-dotenv
requests
matplotlib
flake8
pylint
pytest
